+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add category filtering when adding mods/shaders/worlds/resource packs
 - List the timeout and concurrent connection settings when downloads fail
 - Remove initial memory argument [#727]
+- Add name of platform an instance is from when launching [#740]
 
 ### Fixes
 - Remove old OmitStackTraceInFastThrow JVM arg